LayoutTests/inspector/page/overrideUserAgent-expected.txt

Tests for the Page.overrideUserAgent command.


== Running test suite: Page.overrideUserAgent
-- Running test case: Page.overrideUserAgent.Custom
PASS: There should be a default UserAgent.
Overriding UserAgent...
Reloading...
PASS: The UserAgent should have changed.
Removing UserAgent override...
Reloading...
PASS: The UserAgent should be back to its default value.
